+++ src/app/dashboard/dashboard.component.ts
@@ -55,15 +55,16 @@ export class DashboardComponent implements OnInit {
   }
 
   public removeLoginLocation(id) {
-    let data = {
-      id: id,
-      user_id: this.user_id
-    }
-    this.dashboardService.removeLoginLocation(data).subscribe(
-      data => {
-        this.userLocations = data;
+    if(confirm('Are you sure to Remove this Locaton?')){
+      let data = {
+        id: id,
+        user_id: this.user_id
       }
-    );
+      this.dashboardService.removeLoginLocation(data).subscribe(
+        data => {
+          this.userLocations = data;
+        }
+      );
+    }
   }
-
 }
